[ initial ] Add workshop project and notes
Speaker went _way_ too fast and had too much stuff to cover: this was
not as much a workshop as it was an info-dump. There were some useful
things along the way, and I got an overview of the general structure,
language, and ideas, but no way would one have been able to follow
along; bringing a laptop with Flutter set up was pointless. Essentially,
there was no guidance for what to do (and what there was was sped
through extremely quickly). Use a smaller example next time, or use the
whole session for the workshop and don't try to also fit a talk in
there.